requirements.txt
Django==1.8
django-extensions==1.5.2
django-haystack==2.3.1
django-static-precompiler==0.9
docopt==0.4.0
elasticsearch==1.4.0
futures==2.2.0
gnureadline==6.3.3
ipdb==0.8
ipython==3.0.0
mailchimp==2.0.9
Markdown==2.6.1
Pillow==2.7.0
python-ptrace==0.8.1
requests==2.6.2
six==1.9.0
urllib3==1.10.4
